Workaround to issue 560 - Z fighting in c310 panel.
Underlying problem appears to be that depth-test=true causes
z-buffer fighting for 2D panels that are placed within the 3D
object.

Workaround is to remove the depth test. This has a knock-on effect
of meaning that the yokes hide behind the panel, so the panel
background has been removed, and the magnetos moved to a less
obtrusive place.